Fix the wrong line height in list.
I find the line is too high if some words are bold(or having some style). So try
to update the hextra theme to fix that. But after I did that, hugo failed to
build the site.

After analyzing the error code, I realize the reason may be that some varibles
in hugo are out of date. So I update the hugo binary by downloading the latest
release from github.

Luckily, everything works well.